+++ CHANGES.rst
@@ -10,6 +10,8 @@ https://github.com/zopefoundation/Zope/blob/4.x/CHANGES.rst
 5.8.7 (unreleased)
 ------------------
 
+- Separate ZODB connection information into new ZODB Connections view
+
 - Update the Ace editor in the ZMI.
 
 - Restrict access to static ZMI resources.